[backtrace] using one provided by context
Only Qt can get enough information about context so I do see a way to
make backtrace more verbose only using information provided by
context. So, switching to the use of the context backtrace list. To make
it looking in a way compatible with compilation modes it can be parsed
and modified

Signed-off-by: Denis Zalevskiy <denis.zalevskiy@jollamobile.com>