[cutes] global object is frozen in QQmlEngine
So, native libraries are loaded as an object, "cutes" is set in
QQmlContext for QQmlEngine, separate engine is run to load pure js
modules to allow 'em to access "cutes". I hope there will not be more
pitfalls :)

TODO: in qml it is possible to use CutesAdapter to access cutes
environment w/o directly setting context property. It should be more
portable

Signed-off-by: Denis Zalevskiy <denis.zalevskiy@jolla.com>